# src/infra/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação Swagger do prjeto de registro de forms",
        "version": "1.0.0",
        "title": "registro_form",
        "contact": {
            "email": "keltonmof@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/formulario": {
            "post": {
                "summary": "Registra o usuário no arquivo",
                "tags": ["Formulário"],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Resposta",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postGravarFormulario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jogador Registrado com Sucesso",
                        "schema": {
                            "$ref": "#/definitions/responseGravarFormulario"
                        }
                    },
                    "400": {
                        "description": "Erro ao processar requisição",
                        "schema": {
                            "$ref": "#/definitions/responsePadrao"
                        }
                    },
                    "422": {
                        "description": "Erro ao processar nome do usuário",
                        "schema": {
                            "$ref": "#/definitions/responsePadrao"
                        }
                    }
                }
            },
            "get": {
                "summary": "Valida se pode começar a jogar",
                "tags": ["Formulário"],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "Formulario",
                        "description": "Nome do formulário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorno correto da validação",
                        "schema": {
                            "$ref": "#/definitions/getRetornarFormularioResponse"
                        }
                    },
                    "400": {
                        "description": "Erro ao processar requisição",
                        "schema": {
                            "$ref": "#/definitions/responsePadrao"
                        }
                    },
                    "422": {
                        "description": "Erro ao processar nome do usuário",
                        "schema": {
                            "$ref": "#/definitions/responsePadrao"
                        }
                    }
                }
            }
        },
        "/reset": {
            "put": {
                "summary": "Reinicia o arquivo",
                "tags": ["Formulário"],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Resposta",
                        "required": true,
                        "schema": {
                            "type":"object",
                            "required": ["Formulario"],
                            "properties": {
                                "Formulario": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorno correto da validação",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "type": "integer"
                                },
                                "resetado": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "responsePadrao": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responseGravarFormulario": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "gravado": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "postGravarFormulario": {
            "type": "object",
            "required": ["Nome", "Telefone", "Email", "Formulario"],
            "properties": {
                "Nome": {
                    "type": "string"
                },
                "Telefone": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Formulario": {
                    "type": "string"
                }
            }
        },
        "getRetornarFormularioResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "arquivo": {
                    "type": "string"
                },
                "nomeArquivo": {
                    "type": "string"
                },
                "encoding": {
                    "type": "string"
                }
            }
        }
    }
}